Add partially impl. glass container block entity
This block initially has an IngredientGroup which serves as its contents, and can accept an unlimited amount of items and fluids for now. Temporary
functionality is put in to demonstrate recognition of the GlassContainer's contents; in particular the GlassContainerBlock will drop its corresponding BlockEntity's contents if it contains only ItemStacks and no FluidInstances.