Update to 19w12b, add revolver item w/ .obj model
Revolver model assets were prepared primarily for use with the Architect model loading library. However, the most current version of the library is being refactored to better optimize .obj loading, so it is currently unimplemented.

I built a temporary patched version that reimplemented the old but inefficient .obj loading mechanism. Seeing that this will eventually be phased out for a properly updated version of Architect that can handle .obj model loading and distribution of the jar would require me to make my hacky fork visible source, I decided not to commit the modified jar. Hopefully, this will very likely be resolved before I push out a proper release of Alchym.