Implement basic 'TransmutationInterface's
This class will serve as the key object to help tackle #4. A TransmutationInterface
aims to expose some kind of generic pushing/pulling functionality to an endpoint that
can receive or transmit items. In the context of the issue,
TransmutationInterfaces will act as the "source" and "target" groups of
a TransmutationAction.

Two implementations of TransmutationInterfaces are created, though they
aren't fully implemented out:
- DryTransmutationInterface, to modify ItemEntities within the World
- WetTransmutationInterface, to modify Ingredients within the
SolutionGroup of a GlassContainerBlockEntity