Make book page initializer use lazy initialization
Initialization of the page tree will be pushed from the init stage of
the mod to when the Alchymic Reference is opened for the first time.
This ensures that MinecraftClient#textRenderer is non-null upon
initialization, which is necessary for the page tree to be created.

This commit is only required until the Fabric API includes a
post-initialization hook, which would be guaranteed to execute
after the graphical client window has been created.